* Pass the strutil and filesystem unit tests.
* Correct default font for oiiotool --text.
* Enable the two digit exponent format earlier in oiiotool, allowing more plugin tests to pass.
* Fix some down-casting warning fixes.
* Added the build option CMAKE_USE_FOLDERS to organize the solution projects in folders.

+++ src/oiiotool/oiiotool.h
@@ -91,7 +91,7 @@ public:
     Timer total_readtime;
     Timer total_writetime;
     double total_imagecache_readtime;
-    typedef std::map<std::string, float> TimingMap;
+    typedef std::map<std::string, double> TimingMap;
     TimingMap function_times;
     bool enable_function_timing;
